+++ contrib/toolkits/xgks/lib/xgks/transforms.c
@@ -686,17 +686,17 @@ fprintf(stderr, "gevaltran:transforms.c: input result = %f %f %f %f %f %f\n",
                 ndc_ref = *ppoint;
                 ndc_shift = *pshift;
         }
-        c = (double)(cos( (double) angle) );    /* c1175 */
-        s = (double)(sin( (double) angle) );    /* c1175 */
+        c = (cos( (double) angle) );    /* c1175 */
+        s = (sin( (double) angle) );    /* c1175 */
 
 #ifdef TRANDEBUG
 fprintf(stderr, "gevaltran:transforms.c: (cosine angle)c=%f, (sine angle)s=%f\n", c, s);
 #endif
 
-        result[0][0] = (Gfloat)(pscale->x * (double)c);         /* c1175 */
-        result[0][1] = (Gfloat)(pscale->x * (double)s);         /* c1175 */
-        result[1][0] = (Gfloat)(-(pscale->y * (double)s));      /* c1175 */
-        result[1][1] = (Gfloat)(pscale->y * (double)c);         /* c1175 */
+        result[0][0] = (Gfloat)(pscale->x * c);         /* c1175 */
+        result[0][1] = (Gfloat)(pscale->x * s);         /* c1175 */
+        result[1][0] = (Gfloat)(-(pscale->y * s));      /* c1175 */
+        result[1][1] = (Gfloat)(pscale->y * c);         /* c1175 */
         result[0][2] = (ndc_ref.x+ndc_shift.x) - (ndc_ref.x*result[0][0]) - (ndc_ref.y*result[1][0]);
         result[1][2] = (ndc_ref.y+ndc_shift.y) - (ndc_ref.x*result[0][1]) - (ndc_ref.y*result[1][1]);
 #ifdef TRANDEBUG
@@ -746,8 +746,8 @@ Gint gaccumtran(Gfloat segtran[2][3], Gpoint *ppoint, Gpoint *pshift, Gfloat ang
                 ndc_shift = *pshift;
         }
         /* calculate cosine and sine ONCE */
-        c = (double)cos((double)angle);         /* c1175 */
-        s = (double)sin((double)angle);         /* c1175 */
+        c = cos((double)angle);         /* c1175 */
+        s = sin((double)angle);         /* c1175 */
 
         result[0][0] = (segtran[0][0]*pscale->x*c) + (segtran[1][0]*pscale->x*s);       /* c1175 */
         result[0][1] = (segtran[0][1]*pscale->x*c) + (segtran[1][1]*pscale->x*s);       /* c1175 */
